recheck and change path to index
